feat(builtin)!: add built-in command handling and cd command
Added initial support for handling shell built-in commands, starting
with the `cd` command. This allows users to change directories within
the shell environment.

For better organization, header files have been moved to the `include`
folder, making them easier to import in built-in command files, while
C language files remain in their original location.

Note: This is a preliminary test, and only the `cd` command is
functional at this stage.

Refs: #10
Co-authored-by: Kim Hyunseo <rickroot30@gmail.com>